requirements.txt
requests
black
smolagents[openai]
python-dotenv
langchain_community
rank_bm25
